refactore: resultCenter
